Add waze project documents
The project objective was to Investigate and understand the data provided regarding Waze user’s behavior. This data is a data extract that is planned to be used in a retention model to improve user retention.

The results are posted in an Executive Summary and also in a Jupyter Notebook with the Python Code developed.